Add files via upload
Problem Overview and Objective

Business Description
INN Hotels Group has a chain of hotels in Portugal, they are facing problems with high number of booking cancellations. A significant number of bookings are called off as well as no-shows. The typical reasons include changed plans, scheduling conflicts, as well as customer specific behavior. This is made easier by the option to cancel for free or with little penalty. The cancellation of bookings negatively impact: 1. revenue when the hotel cannot resell the room. 2. marketing costs by increasing commissions paid for publicity to help sell these rooms. 3. profitability because last minute price reductions are necessary to sell the room in short notice. 4. personnel costs because extra hours are required to service guests.

Objective
To develop a Machine Learning solution that can help predict cancellations. Identifying them in advance will help in implementing contingency and cancellation plans.